oh boy!
Need to add a div in the middle of my structure and since I decided to challenge myself and use 0 classes and 0 IDs(oh god why) it was a real workout for my brain.

anyway I finished the tablet and desktop views atleast the general structure of them.

now I just need to style the website, add fonts, make the colors nice and add some texts.