transition fix
The background-image transitions I used only work in google chrome and not in Firefox. Changed the structure to have a background and a <img> and just use opacity to change it

the transition does acutally look nicer like this imo since it doesn't need to fully complete to not clip forward in the animation.